requirements.txt
zeep==4.1.0
python-dotenv==0.19.1
pydantic==1.8.2